complete code

+++ Backend/src/main/java/com/Ecommerce_Multivendor/Backend/repository/TransactionRepository.java
@@ -1,12 +1,11 @@
 package com.Ecommerce_Multivendor.Backend.repository;
 
-
 import com.Ecommerce_Multivendor.Backend.model.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 
     List<Transaction> findBySellerId(Long sellerId);
 }
